Implementation bdd avec marker et derniere position

+++ app/src/main/java/com/example/sanoyan/googlemaps/bdd/Position.java
@@ -20,11 +20,11 @@ public class Position {
         this.utilisateur = utilisateur;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
